+++ Assets/Scripts/PGApi/KenPerlin.cs
@@ -2,15 +2,15 @@
 {
 	public static class Perlin
 	{
-        public static double[,] Generate2DMapUnity(int width, int height, float scale, float offsetX, float offsetY)
+        public static float[,] Generate2DMapUnity(int width, int height, float scale, float offsetX, float offsetY)
 		{
-			double[,] noiseMap = new double[width, height];
+			float[,] noiseMap = new float[width, height];
 			for (int y = 0; y < height; y++)
 			{
 				for (int x = 0; x < width; x++)
 				{
-					double xCoord = (double)x / width * scale + offsetX;
-					double yCoord = (double)y / height * scale + offsetY;
+					float xCoord = (float)x / width * scale + offsetX;
+					float yCoord = (float)y / height * scale + offsetY;
 					noiseMap[x, y] = PGApi.PerlinNoise_Improved2D(xCoord, yCoord);
 				}
 			}
